// lib/api/generated/model/createBrandCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface CreateBrandCommand {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  link?: string | null;
  sortOrder?: number;
}