// lib/api/generated/model/updateBrandCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface UpdateBrandCommand {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  link?: string | null;
  sortOrder?: number;
}